chore: extract react components
