+++ src/feature/area-chart/area-chart.tsx
@@ -5,9 +5,9 @@ import { NonIdealState } from "../../components/non-ideal-state";
 import { defaultPriceFormat } from "../../helpers";
 import { useThrottledResizeObserver } from "../../hooks";
 import { ThemeVariant } from "../../types";
+import { Tooltip, TooltipProps } from "../../ui/components";
 import styles from "./area-chart.module.css";
 import { Chart } from "./chart";
-import { Series, Tooltip, TooltipProps } from "./components";
 import { getColors } from "./helpers";
 
 /**
