--- src/feature/line-chart/components/custom-tooltip/index.ts
@@ -1,1 +1,0 @@
-export * from "./custom-tooltip";
