--- src/feature/line-chart/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./custom-tooltip";
-export * from "./tooltip";
