--- src/feature/price-chart/components/tooltip/index.ts
@@ -1,1 +1,0 @@
-export * from "./tooltip";
