+++ src/feature/price-chart/contents.ts
@@ -1,9 +1,8 @@
-import { range } from "../../helpers";
 import { Renderer } from "../../renderer";
 import { Container } from "../../renderer/display";
 import { ScaleLinear, ScaleTime } from "../../types";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "./constants";
-import { Area, HorizontalGrid, VerticalGrid } from "./display-objects";
+import { Area, HorizontalGrid, VerticalGrid } from "../../ui/display-objects";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
 import { Colors } from "./helpers";
 
 type ContentsColors = Pick<
